rename sync operation, and remove `cancel` job operation
