+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/_operations/__init__.py
@@ -7,14 +7,12 @@
 # --------------------------------------------------------------------------
 
 from ._operations import ConversationAnalysisClientOperationsMixin
-from ._operations import AnalyzeConversationOperations
 
 from ._patch import __all__ as _patch_all
 from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
 from ._patch import patch_sdk as _patch_sdk
 __all__ = [
     'ConversationAnalysisClientOperationsMixin',
-    'AnalyzeConversationOperations',
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])
 _patch_sdk()
