+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/_vendor.py
@@ -17,18 +17,6 @@ if TYPE_CHECKING:
     from azure.core import PipelineClient
 
 
-def _format_url_section(template, **kwargs):
-    components = template.split("/")
-    while components:
-        try:
-            return template.format(**kwargs)
-        except KeyError as key:
-            formatted_components = template.split("/")
-            components = [
-                c for c in formatted_components if "{}".format(key.args[0]) not in c
-            ]
-            template = "/".join(components)
-
 class MixinABC(ABC):
     """DO NOT use this class. It is for internal typing use only."""
     _client: "PipelineClient"
