+++ sdk/cognitivelanguage/azure-ai-language-conversations/samples/async/sample_analyze_orchestration_app_conv_response_async.py
@@ -42,7 +42,7 @@ async def sample_analyze_orchestration_app_conv_response_async():
     client = ConversationAnalysisClient(clu_endpoint, AzureKeyCredential(clu_key))
     async with client:
         query = "Send an email to Carol about the tomorrow's demo"
-        result = await client.analyze_conversations(
+        result = await client.analyze_conversation(
             task={
                 "kind": "Conversation",
                 "analysisInput": {
