+++ sdk/cognitivelanguage/azure-ai-language-conversations/samples/async/sample_analyze_orchestration_app_luis_response_async.py
@@ -42,7 +42,7 @@ async def sample_analyze_orchestration_app_luis_response_async():
     client = ConversationAnalysisClient(clu_endpoint, AzureKeyCredential(clu_key))
     async with client:
         query = "Reserve a table for 2 at the Italian restaurant"
-        result = await client.analyze_conversations(
+        result = await client.analyze_conversation(
             task={
                 "kind": "Conversation",
                 "analysisInput": {
