+++ sdk/cognitivelanguage/azure-ai-language-conversations/samples/async/sample_analyze_orchestration_direct_target_async.py
@@ -44,7 +44,7 @@ async def sample_analyze_orchestration_direct_target_async():
     async with client:
         query = "How are you?"
         qna_app = "ChitChat-QnA"
-        result = await client.analyze_conversations(
+        result = await client.analyze_conversation(
             task={
                 "kind": "Conversation",
                 "analysisInput": {
