+++ sdk/cognitivelanguage/azure-ai-language-conversations/tests/async/test_conversation_app_async.py
@@ -23,7 +23,7 @@ class ConversationAppAsyncTests(AsyncConversationTest):
         client = ConversationAnalysisClient(endpoint, AzureKeyCredential(key))
         async with client:
             query = "Send an email to Carol about the tomorrow's demo"
-            result = await client.analyze_conversations(
+            result = await client.analyze_conversation(
                 task={
                     "kind": "Conversation",
                     "analysisInput": {
