+++ sdk/cognitivelanguage/azure-ai-language-conversations/tests/async/test_orchestration_app_qna_response_async.py
@@ -22,7 +22,7 @@ class OrchestrationAppQnaResponseAsyncTests(AsyncConversationTest):
         client = ConversationAnalysisClient(endpoint, AzureKeyCredential(key))
         async with client:
             query = "How are you?"
-            result = await client.analyze_conversations(
+            result = await client.analyze_conversation(
                 task={
                     "kind": "Conversation",
                     "analysisInput": {
