+++ sdk/cognitivelanguage/azure-ai-language-conversations/tests/test_conversation_app.py
@@ -23,7 +23,7 @@ class ConversationAppTests(ConversationTest):
         client = ConversationAnalysisClient(endpoint, AzureKeyCredential(key))
         with client:
             query = "Send an email to Carol about the tomorrow's demo"
-            result = client.analyze_conversations(
+            result = client.analyze_conversation(
                 task={
                     "kind": "Conversation",
                     "analysisInput": {
