+++ sdk/cognitivelanguage/azure-ai-language-conversations/tests/test_orchestration_app_qna_response.py
@@ -25,7 +25,7 @@ class OrchestrationAppQnaResponseTests(ConversationTest):
         client = ConversationAnalysisClient(endpoint, AzureKeyCredential(key))
         with client:
             query = "How are you?"
-            result = client.analyze_conversations(
+            result = client.analyze_conversation(
                 task={
                     "kind": "Conversation",
                     "analysisInput": {
